% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDashBoardDataRow.R
\name{getDashBoardDataRow}
\alias{getDashBoardDataRow}
\title{Return a row from data table based on current time.}
\usage{
getDashBoardDataRow()
}
\value{
A row of data from dashBoardData
}
\description{
Assumes the presence of dashBoardData
}
\examples{
getDashBoardDataRow()
}
