% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pulseToengine.R
\name{plot_pulseToengine}
\alias{plot_pulseToengine}
\title{plot pulsometer against engine efficiency}
\usage{
plot_pulseToengine(pulsometer_10, engine_10)
}
\arguments{
\item{pulsometer_10}{a vector of 10 values from pulsometer.}

\item{engine_10}{a vector of 10 values from engine efficiency.}
}
\value{
a plot
}
\description{
plot pulsometer against engine efficiency
}
\examples{
plot_pulseToengine(dashBoardData[1:10, "Pulsometer_readout"],  dashBoardData[1:10, "Engine_efficiency"])
}
