% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pulseToengine.R
\name{plot_pulseToengine}
\alias{plot_pulseToengine}
\title{plot pulsometer against engine efficiency}
\usage{
plot_pulseToengine(howManyToPlot = 10)
}
\arguments{
\item{howManyToPlot}{how many points to plot. Default 10}
}
\value{
a plot with splines
}
\description{
plot pulsometer against engine efficiency
}
\examples{
plot_pulseToengine()
}
